--- a/Dockerfile.server
+++ b/Dockerfile.server
@@ -1,0 +1,13 @@
+FROM node:24-alpine
+WORKDIR /app
+
+COPY package.json package-lock.json ./
+RUN npm ci --omit=dev --legacy-peer-deps
+COPY . .
+
+RUN npm run build
+
+EXPOSE 8080
+
+ENV DIST=1
+CMD ["npm", "start"]