--- a/gulp/update-vendor.mjs
+++ b/gulp/update-vendor.mjs
@@ -13,6 +13,12 @@ const vendorFiles = [
 	'./node_modules/swiped-events/src/swiped-events.js',
 ];
 
+// Special handling for metar-taf-parser - only copy main file and English locale
+const metarFiles = [
+	'./node_modules/metar-taf-parser/metar-taf-parser.js',
+	'./node_modules/metar-taf-parser/locale/en.js',
+];
+
 const copy = () => src(vendorFiles)
 	.pipe(rename((path) => {
 		path.dirname = path.dirname.toLowerCase();
@@ -22,6 +28,14 @@ const copy = () => src(vendorFiles)
 	}))
 	.pipe(dest('./server/scripts/vendor/auto'));
 
-const updateVendor = series(clean, copy);
+const copyMetar = () => src(metarFiles, { base: './node_modules/metar-taf-parser' })
+	.pipe(rename((path) => {
+		path.basename = path.basename.toLowerCase();
+		path.extname = path.extname.toLowerCase();
+		if (path.basename === 'metar-taf-parser') path.extname = '.mjs';
+	}))
+	.pipe(dest('./server/scripts/vendor/auto'));
+
+const updateVendor = series(clean, copy, copyMetar);
 
 export default updateVendor;