--- a/server/scripts/modules/autocomplete.mjs
+++ b/server/scripts/modules/autocomplete.mjs
@@ -192,7 +192,7 @@ class AutoComplete {
 
 		let result = this.cachedResponses[search];
 		if (!result) {
-			// make the request
+			// make the request; using json here instead of safeJson is fine because it's infrequent and user-initiated
 			const resultRaw = await json(url);
 
 			// use the provided parser