## Summary

This PR introduces comprehensive improvements to caching, error handling, performance, and user experience, including:

- Complete server-side caching proxy replacing CORS workarounds
- Enhanced kiosk mode and scanline rendering
- METAR data augmentation for missing observations
- Content-aware timing and hardware-accelerated scrolling
- Centralized error handling across all modules

## Background

Hi Matt.

I found out about this project recently and was immediately enamored! (I grew up in a household that had TWC on almost any time it wasn't "prime time"!) My dad's birthday was coming up and I knew this would make the cornerstone of the perfect (and unique) gift. So I found a 10" 4:3 LCD monitor with HDMI input on Amazon, ordered a 2GB Raspberry Pi 4, configured it to auto-boot into kiosk mode, and then designed and 3D-printed a case to mount the Pi to the VESA mount on the back of the monitor. (He _loved_ it. 🙂)

While I was tinkering with getting it set up, I started looking in to some occasional "strangeness" with the weather data that I saw in the JavaScript console. One thing led to another, and I had implemented an enhanced proxy solution to provide as much insight as possible into what was happening with the various API calls (and more importantly, _how often_). Then it was adding full _cache_ support to the proxy to ensure I had insight/control into what was actually happening with caching _without_ the browser being involved. Below is an example of the server-side logging showing cache hits, adds, stale cache entries that haven't changed (so there's no need to re-request the full resource), etc. Also notice that the radar images do _not_ have cache headers, but I heuristically cache them as allowed by RFC 7234. (All this makes WS4KP a _much_ nicer "netizen" in terms of API calls, especially on the Mesonet servers! 🎉)
```
🎯 Hit     | https://www.spc.noaa.gov/products/outlook/day2otlk_cat.nolyr.geojson (age: 1199s, remaining: 2401s)
🎯 Hit     | https://www.spc.noaa.gov/products/outlook/day1otlk_cat.nolyr.geojson (age: 1199s, remaining: 2401s)
🎯 Hit     | https://www.spc.noaa.gov/products/outlook/day3otlk_cat.nolyr.geojson (age: 1199s, remaining: 2401s)
📤 Sent    | https://mesonet.agron.iastate.edu/archive/data/2025/06/22/GIS/uscomp/?F=0&P=n0r*.png (no cache directives; not cached)
🎯 Hit     | https://mesonet.agron.iastate.edu/archive/data/2025/06/22/GIS/uscomp/n0r_202506221545.png (age: 1199s, remaining: 2401s)
🎯 Hit     | https://mesonet.agron.iastate.edu/archive/data/2025/06/22/GIS/uscomp/n0r_202506221550.png (age: 1199s, remaining: 2401s)
🎯 Hit     | https://mesonet.agron.iastate.edu/archive/data/2025/06/22/GIS/uscomp/n0r_202506221555.png (age: 600s, remaining: 3000s)
🎯 Hit     | https://mesonet.agron.iastate.edu/archive/data/2025/06/22/GIS/uscomp/n0r_202506221600.png (age: 599s, remaining: 3001s)
🌐 Add     | https://api.weather.gov/gridpoints/IND/83,92/stations?limit=1 (explicit 120s TTL, expires: 2025-06-22T16:15:40.570Z)
🔄 Unchg   | https://api.weather.gov/stations/KMKE/observations/latest (got 304 Not Modified; refreshing cache expiry by 300s)
🌐 Add     | https://mesonet.agron.iastate.edu/archive/data/2025/06/22/GIS/uscomp/n0r_202506221610.png (heuristic 3600s TTL, expires: 2025-06-22T17:13:40.656Z)
```

After that, it snowballed in to more and more changes as I ran into the occasional error, or thought _"it might be cool if..."_

## Details of Changes

- Enhance the kiosk mode startup experience so it is more "appliance like" (so when my dad plugged it in on his birthday, it looked "professional")
- Spent a bunch of time optimizing the _scanlines_ feature since I think it adds a lot to the presentation. (The original implementation at 1024x768 resulted in pretty bad banding, _click to enlarge_.)
<table>
<tr>
<td width="50%">

**Before**
![Before](https://github.com/user-attachments/assets/fa1546b0-1146-45f0-a920-af56995f1162)

</td>
<td width="50%">

**After** 
![After](https://github.com/user-attachments/assets/0d4c8975-6eba-4609-9bfc-a3259a14a75e)

</td>
</tr>
</table>

- Wanted to customize the "travel forecast" with cities important to my dad. I didn't like the complexity of manually creating the `.js` versions, so found a way to use the `.json` version
- Wanted a way to dynamically enable/disable debug logs, so added a module for that
- Needed a way to see details of transition times, so added logging to `weatherdisplay.mjs`
- Saw that I could make some observations "valid" by augmenting them with data present in the METAR, so added the dependency-free `metar-taf-parser` module to the build
- I added a utility function to augment missing weather observation data from the METAR
- Then updated `currentweather.mjs` and `latestobservations.mjs` to augment their from the METAR when possible
- Started moving modules to use centralized error handling so all network errors would be handled consistently, only "bubbling up" unexpected errors
- Added a module to remove expired forecast periods from "stale" API responses
- Updated `extendedforecast.mjs` to skip expired forecasts, changed it to focus on full days, and fixed a day/night temperature pairing issue
- Updated `localforecast.mjs`. In addition to skipping expired forecasts, I added a content-aware transition timing system that uses a DOM-based measurement system to determine the number of lines in each forecast and adjust the dwell time depending on how "full" the screen is.
- Improved the processing and error handling in `spc-outlook.mjs`
- Improved error handling in `regionalforecast.mjs` and `utils/weather.mjs`
- Worked on the weather radar files: optimized radar API usage by only fetching yesterday's data when needed (the Apache index file does not any have cache headers so it's never cached), use centralized URL rewriting for caching proxy support, added debug logging throughout radar processing pipeline, improve canvas context validation and error recovery, and tried to handle worker errors gracefully (by skipping the animation)
- Starting looking into scroll performance and flexibility. I added `utils/scroll-timing.mjs` and worked to replace "magic numbers" with seconds-based timing constants. Switched from `scrollTo` to hardware-accelerated `transform`. Added scroll caching to prevent repeated  DOM queries (which can trigger layout recalculations each time these properties are accessed) on every scroll cycle. Fixed calculations to allow an _arbitrarily-sized_ list of travel forecast cities instead of "20".
- Did the same thing to `hourly.mjs`, including updating the calculations to support flexible hourly forecast lengths (in the future, there could be a setting to display every-other-hour or something)
- Worked on alerts ("hazards") next (both `hazards.mjs` and `currentweatherscroll.mjs`). Updated timing to use seconds-based timing constants. Switched from `scrollTo` to `transform` and did the scroll-caching thing. Updated the CSS for the horizontal alert scroll to go edge-to-edge. Integrated the global speed settings into horizontal scroll timing. 

There are also some minor housekeeping changes: updating the VS Code settings, adding start/stop commands and updating lint rules in `package.json`, etc.

I realize this is an absolutely **massive** set of changes, but I wanted to share them because I think this project is terrific and wanted to give back. With that being said, I started working from the v5.23.3 tag, not realizing how active this project was. I'm not sure what would be involved in applying all the updates since that tag to this PR, because I didn't want to go down that path if you had no interest in merging this. And if you're not, I completely understand — I'll just use them in my private fork. But if you are, it might be enough to bump up to v6.

_In any case, thanks for WS4KP!_

P.S. next up I'm looking in to adding support to fallback to the legacy `MapClick.php` interface when there's stale observation data from the API since I regularly see the API returning observation data that is not current, or forecast data that has expired periods. (Hence the METAR augmentation and expired forecast filtering features...) Is that just me?!