@eddyg 2025-06-23
Thanks for taking a look, Matt.

Happy to discuss any "design decisions" with you if you have questions about why I did something.

I had two over-arching goals when I started on this:

1. A "clean" Console while the app was running. Meaning, if the API was working "as intended", there wouldn't be any messages. When there was "anomalous" behavior, I do want to see stuff like this:
```
[Log] Latest Observations for station XXXX were augmented with METAR data for windSpeed, windDirection
[Log] Forecast: Removed expired periods: This Afternoon (ended 2025-06-22T18:00:00-04:00), Tonight (ended 2025-06-23T06:00:00-04:00)
[Log] Latest Observations for station XXXX were augmented with METAR data for temperature
```

2) a caching proxy log that showed that each API request was only put on-the-wire when necessary (I think it's important to be a good "netizen" when it comes to API use).

One problem I ran into (with Safari in particular) was aggressive caching of API responses. I could see in the Network tab that the browser was _not_ sending out new API requests (even though the cache headers indicate a 300s TTL) but instead was returning results from the cache. By being able to control the cache headers in the API responses sent to the browser, I alleviated that problem and got a much better picture of how often API requests were coming from the app.

Thanks again!