@rmitchellscott 2025-06-25
This seems to break static builds and therefore the project's current Docker approach:
```
ws.min.js?_=5.25.3:1  ❌ Fetch failed for https://api.weather.gov/points/39.3615,-104.6020 (Fetch error 404 Not Found while fetching http://localhost:8080/api/points/39.3615,-104.6020)
```