@rmitchellscott 2025-06-25
@eddyg Breaking static builds seems like a poor direction to go in. There's no hard reason (to my knowledge) for this project to require a running Node backend. For example, the previous Node docker image was 362MB, while the static version is 68MB. You could also run the static site from an S3 bucket, etc. 

This patch got your codebase running statically for me:

```diff
diff --git a/server/scripts/modules/utils/url-rewrite.mjs b/server/scripts/modules/utils/url-rewrite.mjs
  index 949b33a..5e51ead 100644
  --- a/server/scripts/modules/utils/url-rewrite.mjs
  +++ b/server/scripts/modules/utils/url-rewrite.mjs
  @@ -8,6 +8,14 @@ const rewriteUrl = (_url) => {
   	// Handle both string URLs and URL objects
   	const url = typeof _url === 'string' ? new URL(_url) : new URL(_url.toString());

+	// Skip rewriting in static builds (no proxy server available)
+	const isStaticBuild = document.querySelector('script[src*="ws.min.js"]') !== null;
+
+	if (isStaticBuild) {
+		return url;
+	}
+
   	// Rewrite the origin to use local proxy server
   	if (url.origin === 'https://api.weather.gov') {
   		url.protocol = window.location.protocol;

```

Curious if something like this would work to maintain compatibility with both methods of running the project.