@eddyg 2025-06-25
I've updated my latest commit with `Dockerfile` and `Dockerfile.static`. I also adjusted the logic that adds the `User-Agent` header to avoid CORS issues in "static" mode when `User-Agent` is present.