@rmitchellscott 2025-06-25
A couple breaking changes with this:
1. `RUN rm -f dist/playlist.json` in the regular Dockerfile breaks the music, that's a workaround specifically for static deployments. 
2. Removing `add_header X-Weatherstar true always;` from nginx.conf breaks music in the static deployment.

I wonder if it'd be better to invert the logic on the `window.WS4KP_STATIC_BUILD` so that the dynamic backend injects `window.WS4KP_SERVER_AVAILABLE` as opposed to requiring users who want to host statically manually change HTML, something like this maybe:

```diff
diff --git a/index.mjs b/index.mjs
index ddabd75..f74d1a7 100644
--- a/index.mjs
+++ b/index.mjs
@@ -131,9 +131,34 @@ if (process.env?.DIST === '1') {
        // 'npm run build' and then 'DIST=1 npm start'
        app.use('/scripts', express.static('./server/scripts', staticOptions));
        app.use('/geoip', geoip);
+       // Serve index.html with server availability flag injected
+       app.get('/', (req, res) => {
+               let html = fs.readFileSync('./dist/index.html', 'utf8');
+
+               // Inject server availability flag
+               html = html.replace('<head>', '<head><script>window.WS4KP_SERVER_AVAILABLE=true;</script>');
+
+               res.send(html);
+       });
+
        app.use('/', express.static('./dist', staticOptions));
 } else {
        // Development mode serves files from the server directory: 'npm start'
+       // Inject server availability flag into HTML responses
+       app.use((req, res, next) => {
+               if (req.path === '/' || req.path === '/index.html') {
+                       const originalSend = res.send;
+                       res.send = function(html) {
+                               if (typeof html === 'string' && html.includes('<head>')) {
+                                       html = html.replace('<head>', '<head><script>window.WS4KP_SERVER_AVAILABLE=true;</script>');
+                               }
+                               originalSend.call(this, html);
+                       };
+               }
+               next();
+       });
+
        app.get('/index.html', index);
        app.use('/geoip', geoip);
        app.use('/resources', express.static('./server/scripts/modules'));
diff --git a/server/scripts/modules/utils/fetch.mjs b/server/scripts/modules/utils/fetch.mjs
index 5ba1f8c..cd347c4 100644
--- a/server/scripts/modules/utils/fetch.mjs
+++ b/server/scripts/modules/utils/fetch.mjs
@@ -86,7 +86,7 @@ const fetchAsync = async (_url, responseType, _params = {}) => {
        // User-Agent handling:
        // - Server mode (with caching proxy): Add User-Agent for all requests except excluded hosts
        // - Static mode (direct requests): Only add User-Agent for api.weather.gov, avoiding CORS preflight issues with other services
-       const shouldAddUserAgent = !shouldExcludeUserAgent && (!window.WS4KP_STATIC_BUILD || _url.toString().match(/api\.weather\.gov/));
+       const shouldAddUserAgent = !shouldExcludeUserAgent && (window.WS4KP_SERVER_AVAILABLE || _url.toString().match(/api\.weather\.gov/));
        if (shouldAddUserAgent) {
                headers['user-agent'] = 'Weatherstar 4000+; weatherstar@netbymatt.com';
        }
diff --git a/server/scripts/modules/utils/url-rewrite.mjs b/server/scripts/modules/utils/url-rewrite.mjs
index 17bb7d5..7103720 100644
--- a/server/scripts/modules/utils/url-rewrite.mjs
+++ b/server/scripts/modules/utils/url-rewrite.mjs
@@ -8,8 +8,8 @@ const rewriteUrl = (_url) => {
        // Handle both string URLs and URL objects
        const url = typeof _url === 'string' ? new URL(_url) : new URL(_url.toString());

-       if (window.WS4KP_STATIC_BUILD) {
+       if (!window.WS4KP_SERVER_AVAILABLE) {
                return url;
        }
```