@rmitchellscott 2025-06-25
> Hmm, what if a "server user" (as opposed to a "static user") wants to bind-mount their own music directory? 🤔

Music bind mounts should happen at `/server/music` and if there are tracks there, playlist.json will get dynamically generated, else it falls back to the build-time `dist/playlist.json` which contains the default tracks.

> This header should be present for both deployment methods via the change in index.mjs... is that not working as expected?

This is Express middleware. Express is not running when these files are served statically, there's no Node.js server to execute server-side Javascript, headers have to be set by the web server (nginx in the case of our static Docker setup).